make stuff work a bit more

+++ app/src/main/java/social/spielapp/android/util/DateUtil.java
@@ -4,7 +4,7 @@ import java.text.DateFormat;
 import java.util.Date;
 
 public class DateUtil {
-    public static String timestampToString(int timestamp) {
+    public static String timestampToString(long timestamp) {
         Date date = new Date(timestamp);
         return DateFormat.getDateTimeInstance().format(date);
     }
